Совместимость длины подземных лент с Show Max Underground Distance
Мод печёт индикатор дальности подземников в data-final-fixes из max_distance. Наши длины задавались там же, но позже мода (он грузится раньше по алфавиту), поэтому индикатор отражал длину до правок zzz, а не фактическую.

Установка длин вынесена в data-updates — на стадию раньше; мод читает финал и рисует корректный индикатор сам. Сами длины не меняются.
