+++ mods/zzzparanoidal/data-updates.lua
@@ -2,3 +2,6 @@
 require("tweaks.recipe.angels-components") -- Сюда складывать все рецепты из группы (item-group) = angels-components
 require("tweaks.recipe.angels-power") -- Сюда складывать все рецепты из группы (item-group) = angels-power
 require("tweaks.recipe.production") -- Сюда складывать все рецепты из группы (item-group) = production
+
+-- Длины подземных лент: на data-updates, чтобы show-max-underground-distance (data-final-fixes) читал финал
+require("tweaks.entity.underground-belt-distance")
